remove audio and camera ns from /include
