+++ nc/include/MainCamera.h
@@ -3,7 +3,7 @@
 #include "component/Camera.h"
 #include "component/Transform.h"
 
-namespace nc::camera
+namespace nc
 {
     /** Specifies the camera to use for view and projection calculations.
      *  A valid camera must be set during frame execution. */
