--- nc/include/graphics/TechniqueType.h
@@ -1,12 +1,0 @@
-#pragma once
-
-#include <cstdint>
-
-namespace nc
-{
-    enum class TechniqueType : uint8_t
-    {
-        None = 0,
-        PhongAndUi = 1,
-    };
-}
