+++ nc/source/graphics/resources/CMakeLists.txt
@@ -1,7 +1,6 @@
 target_sources(${ENGINE_LIB}
     PRIVATE
         ${ROOT_DIR}/source/graphics/resources/DepthStencil.cpp
-        ${ROOT_DIR}/source/graphics/resources/Material.cpp
         ${ROOT_DIR}/source/graphics/resources/ImmutableImage.cpp
         ${ROOT_DIR}/source/graphics/resources/ObjectDataManager.cpp
         ${ROOT_DIR}/source/graphics/resources/PointLightManager.cpp
